separated main.rs
